% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ing}
\alias{ing}
\title{ING Data from Sociolinguistic Interviews}
\format{a data frame with 26 columns
\describe{
 \item{Token}{the word the observation was embedded in}
 \item{DepVar}{whether the token was realized with an apical or velar variant}
 \item{Style}{the speaking style}
 \item{GramStatus}{grammatical status of the token}
 \item{Following.Seg}{coding of the following segment}
 \item{Sex}{sex of speaker}
 \item{Age}{decade of age of the speaker}
 \item{Ethnicity}{Ethnicity of the speaker}
}}
\usage{
ing
}
\description{
ING Data from Sociolinguistic Interviews
}
\details{
This (ING) data was coded from sociolinguistic interviews carried out in Philadelphia.
}
\keyword{datasets}

