% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joe_vowels}
\alias{joe_vowels}
\title{Joe's Vowels}
\format{a data frame with 10 columns
\describe{
 \item{year}{year of the interview}
 \item{plt_vclass}{indicates the vowel class}
 \item{word}{the word the vowel was in}
 \item{dur}{duration of the vowel in seconds}
 \item{F1_20, _35, _80}{F1 at 20\%, 35\% and 80\% of the vowel's duration}
 \item{F2_20, _35, _80}{F2a at 20\%, 35\% and 80\% of the vowel's duration}
}}
\usage{
joe_vowels
}
\description{
Joe's Vowels
}
\details{
This vowel data was extracted from a sociolinguistic interview Joe did in 2006
}
\keyword{datasets}

